+++ src/main/java/com/jasonvillar/works/register/service/Service.java
@@ -27,6 +27,7 @@ public class Service {
         this.user = user;
     }
 
-    @ManyToOne
+    @OneToOne
+    @JoinColumn(name = "user_id", referencedColumnName = "id")
     private User user;
 }